NewBackend/requirements.txt
flask
pandas
flask_cors
openai
python_dotenv